[CC] Renaming
