+++ pyleecan/GUI/Dialog/DMachineSetup/DAVDuct/DAVDuct.ui
