+++ pyleecan/GUI/Dialog/DMachineSetup/DAVDuct/PVentCirc/Gen_PVentCirc.py
@@ -1,16 +1,16 @@
 # -*- coding: utf-8 -*-
-"""File generated according to PVentTrap/gen_list.json
+"""File generated according to PVentCirc/gen_list.json
 WARNING! All changes made in this file will be lost!
 """
-from pyleecan.GUI.Dialog.DMachineSetup.SLamParam.DAVDuct.PVentTrap.Ui_PVentTrap import (
-    Ui_PVentTrap,
+from pyleecan.GUI.Dialog.DMachineSetup.DAVDuct.PVentCirc.Ui_PVentCirc import (
+    Ui_PVentCirc,
 )
 
 
-class Gen_PVentTrap(Ui_PVentTrap):
-    def setupUi(self, PVentTrap):
+class Gen_PVentCirc(Ui_PVentCirc):
+    def setupUi(self, PVentCirc):
         """Abstract class to update the widget according to the csv doc"""
-        Ui_PVentTrap.setupUi(self, PVentTrap)
+        Ui_PVentCirc.setupUi(self, PVentCirc)
         # Setup of in_Zh
         txt = self.tr(u"""Number of Hole around the circumference""")
         self.in_Zh.setWhatsThis(txt)
@@ -36,45 +36,23 @@ class Gen_PVentTrap(Ui_PVentTrap):
         self.lf_Alpha0.setToolTip(txt)
 
         # Setup of in_D0
-        txt = self.tr(u"""Hole height""")
+        txt = self.tr(u"""Hole diameters""")
         self.in_D0.setWhatsThis(txt)
         self.in_D0.setToolTip(txt)
 
         # Setup of lf_D0
         self.lf_D0.validator().setBottom(0)
-        txt = self.tr(u"""Hole height""")
+        txt = self.tr(u"""Hole diameters""")
         self.lf_D0.setWhatsThis(txt)
         self.lf_D0.setToolTip(txt)
 
         # Setup of in_H0
-        txt = self.tr(u"""Radius of the hole bottom""")
+        txt = self.tr(u"""Radius of the hole centers""")
         self.in_H0.setWhatsThis(txt)
         self.in_H0.setToolTip(txt)
 
         # Setup of lf_H0
         self.lf_H0.validator().setBottom(0)
-        txt = self.tr(u"""Radius of the hole bottom""")
+        txt = self.tr(u"""Radius of the hole centers""")
         self.lf_H0.setWhatsThis(txt)
         self.lf_H0.setToolTip(txt)
-
-        # Setup of in_W1
-        txt = self.tr(u"""Hole small basis""")
-        self.in_W1.setWhatsThis(txt)
-        self.in_W1.setToolTip(txt)
-
-        # Setup of lf_W1
-        self.lf_W1.validator().setBottom(0)
-        txt = self.tr(u"""Hole small basis""")
-        self.lf_W1.setWhatsThis(txt)
-        self.lf_W1.setToolTip(txt)
-
-        # Setup of in_W2
-        txt = self.tr(u"""Hole large basis""")
-        self.in_W2.setWhatsThis(txt)
-        self.in_W2.setToolTip(txt)
-
-        # Setup of lf_W2
-        self.lf_W2.validator().setBottom(0)
-        txt = self.tr(u"""Hole large basis""")
-        self.lf_W2.setWhatsThis(txt)
-        self.lf_W2.setToolTip(txt)
