+++ pyleecan/GUI/Dialog/DMachineSetup/DAVDuct/PVentCirc/PVentCirc.ui
