+++ pyleecan/GUI/Dialog/DMachineSetup/DAVDuct/PVentCirc/Ui_PVentCirc.py
