+++ pyleecan/GUI/Dialog/DMachineSetup/DAVDuct/PVentPolar/Gen_PVentPolar.py
@@ -2,7 +2,7 @@
 """File generated according to PVentPolar/gen_list.json
 WARNING! All changes made in this file will be lost!
 """
-from pyleecan.GUI.Dialog.DMachineSetup.SLamParam.DAVDuct.PVentPolar.Ui_PVentPolar import (
+from pyleecan.GUI.Dialog.DMachineSetup.DAVDuct.PVentPolar.Ui_PVentPolar import (
     Ui_PVentPolar,
 )
