+++ pyleecan/GUI/Dialog/DMachineSetup/DAVDuct/PVentPolar/PVentPolar.py
