+++ pyleecan/GUI/Dialog/DMachineSetup/DAVDuct/PVentPolar/PVentPolar.ui
