+++ pyleecan/GUI/Dialog/DMachineSetup/DAVDuct/PVentPolar/Ui_PVentPolar.py
