+++ pyleecan/GUI/Dialog/DMachineSetup/DAVDuct/PVentPolar/__init__.py
