+++ pyleecan/GUI/Dialog/DMachineSetup/DAVDuct/PVentTrap/PVentTrap.py
