+++ pyleecan/GUI/Dialog/DMachineSetup/DAVDuct/PVentTrap/Ui_PVentTrap.py
