+++ pyleecan/GUI/Dialog/DMachineSetup/DAVDuct/Ui_DAVDuct.py
