+++ pyleecan/GUI/Dialog/DMachineSetup/DAVDuct/WVent/Ui_WVent.py
