+++ pyleecan/GUI/Dialog/DMachineSetup/DAVDuct/WVent/WVent.ui
